luci-base: luci.js: fix `LuCI.request.poll.add()` exception handling
The poll result callback might cause itself an exception to be thrown. With this change, we differentiate between exceptions thrown due to bad JSON in `responseText` and exception generated inside of the callback itself. The latter is just let through for browser console logging, as poll callbacks must not throw normally.